reorganized the package structure to be like the most popular npm module, like request.
